requirements.txt
torch
transformers>=4.41
accelerate
bitsandbytes
sentence-transformers
faiss-cpu
pypdf
pillow
streamlit
python-dotenv
huggingface_hub
langchain-google-genai